src/main/java/id/ac/ui/cs/advprog/productservice/service/ProductServiceImpl.java
package id.ac.ui.cs.advprog.productservice.service;

import id.ac.ui.cs.advprog.productservice.model.Product;
import id.ac.ui.cs.advprog.productservice.repository.ProductRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.UUID;

@Service
public class ProductServiceImpl implements ProductService {

    @Autowired
    private ProductRepository productRepository;

    @Override
    public Product create(Product product) {
        return productRepository.save(product);
    }

    @Override
    public Product update(Product product) {
        return productRepository.update(product);
    }

    @Override
    public void delete(String id) {
        productRepository.delete(UUID.fromString(id));
    }

    @Override
    public Product findById(String id) {
        return productRepository.findById(UUID.fromString(id));
    }

    @Override
    public List<Product> findAll() {
        return productRepository.findAll();
    }
}
